Add files via upload
This program will display a number menu to the user.
The program will ask the user to select an option from
the menu and to enter a sentence. If they do not put in
a valid option the program will continue to ask them until
they do. To exit the program the user will insert -999. The
user will have 4 options, display the middle character if
it exists, display the sentence in uppercase, display the
sentence in lowercase, and display the sentence backwards.

You are not allowed to use the lolower() method, toupper()
method, vectors, or any method that reverses the sentence
automatically. You are only allowed to have one "return 0"
statement in your program.